add auth verification for reviews

+++ backend/events/EventController.cs
@@ -224,7 +224,6 @@ namespace backend.events
                 return Unauthorized("User is not logged in.");
             }
             
-            
             var eventDto = await eventService.GetDetailedEvent(id);
 
             if (eventDto == null)
@@ -295,13 +294,20 @@ namespace backend.events
         }
 
         // 3. Get all reviews by a specific user
-        [HttpGet("reviews/user/{userId}")]
+        [HttpGet("reviews/user")]
         [ProducesResponseType(typeof(List<ReviewDto>), 200)]
         [ProducesResponseType(500)]
-        public async Task<IActionResult> GetReviewsByUserId(string userId)
+        public async Task<IActionResult> GetReviewsByUserId()
         {
+            JObject? userInfo = HttpContext.Items["UserInfo"] as JObject;
+
+            if (userInfo == null)
+            {
+                return Unauthorized("User is not logged in.");
+            }
             try
             {
+                string userId = userInfo["data"]["user_id"].ToString();
                 var reviews = await eventService.GetAllReviewsByUserIdAsync(userId);
                 return Ok(reviews);
             }
@@ -312,14 +318,21 @@ namespace backend.events
         }
 
         // 4. Get a specific review by user and event
-        [HttpGet("reviews/{userId}/{eventId}")]
+        [HttpGet("reviews/{eventId}")]
         [ProducesResponseType(typeof(ReviewDto), 200)]
         [ProducesResponseType(404)]
         [ProducesResponseType(500)]
-        public async Task<IActionResult> GetReviewOfUserByEvent(string userId, string eventId)
+        public async Task<IActionResult> GetReviewOfUserByEvent(string eventId)
         {
+            JObject? userInfo = HttpContext.Items["UserInfo"] as JObject;
+
+            if (userInfo == null)
+            {
+                return Unauthorized("User is not logged in.");
+            }
             try
             {
+                string userId = userInfo["data"]["user_id"].ToString();
                 var review = await eventService.GetReviewOfUserByEventAsync(userId, eventId);
                 if (review == null)
                 {
@@ -340,6 +353,20 @@ namespace backend.events
         [ProducesResponseType(500)]
         public async Task<IActionResult> CreateReview([FromBody] CreateReviewDto dto)
         {
+            JObject? userInfo = HttpContext.Items["UserInfo"] as JObject;
+
+            if (userInfo == null)
+            {
+                return Unauthorized("User is not logged in.");
+            }
+            
+            string userId = userInfo["data"]["user_id"].ToString();
+
+            if (dto.UserId != userId)
+            {
+                return Unauthorized("User is not authorized.");
+            }
+            
             if (dto == null || string.IsNullOrEmpty(dto.UserId) || string.IsNullOrEmpty(dto.EventId) || string.IsNullOrEmpty(dto.Text) || dto.Score <= 0)
             {
                 return BadRequest(new { error = "Invalid review data." });
@@ -347,7 +374,7 @@ namespace backend.events
 
             try
             {
-                await eventService.CreateReviewAsync(dto.UserId, dto.EventId, dto.Text, dto.Score);
+                await eventService.CreateReviewAsync(dto);
                 return StatusCode(201, new { message = "Review created successfully." });
             }
             catch (KeyNotFoundException ex)
@@ -361,14 +388,22 @@ namespace backend.events
         }
 
         // 6. Delete a review
-        [HttpDelete("reviews/{userId}/{eventId}")]
+        [HttpDelete("reviews/{eventId}")]
         [ProducesResponseType(200)]
         [ProducesResponseType(404)]
         [ProducesResponseType(500)]
-        public async Task<IActionResult> DeleteReview(string userId, string eventId)
+        public async Task<IActionResult> DeleteReview(string eventId)
         {
+            JObject? userInfo = HttpContext.Items["UserInfo"] as JObject;
+
+            if (userInfo == null)
+            {
+                return Unauthorized("User is not logged in.");
+            }
+            
             try
             {
+                string userId = userInfo["data"]["user_id"].ToString();
                 await eventService.DeleteReviewAsync(userId, eventId);
                 return Ok(new { message = "Review deleted successfully." });
             }
@@ -381,8 +416,6 @@ namespace backend.events
                 return StatusCode(500, new { error = ex.Message });
             }
         }
-    
-        
         
         
         // Comments
@@ -421,13 +454,21 @@ namespace backend.events
         }
 
         // 3. Get all comments by a specific user
-        [HttpGet("comments/user/{userId}")]
+        [HttpGet("comments/user")]
         [ProducesResponseType(typeof(List<CommentDto>), 200)]
         [ProducesResponseType(500)]
-        public async Task<IActionResult> GetCommentsByUserId(string userId)
+        public async Task<IActionResult> GetCommentsByUserId()
         {
+            JObject? userInfo = HttpContext.Items["UserInfo"] as JObject;
+
+            if (userInfo == null)
+            {
+                return Unauthorized("User is not logged in.");
+            }
+            
             try
             {
+                string userId = userInfo["data"]["user_id"].ToString();
                 var comments = await eventService.GetAllCommentsByUserIdAsync(userId);
                 return Ok(comments);
             }
@@ -438,14 +479,22 @@ namespace backend.events
         }
 
         // 4. Get specific comments by user and event
-        [HttpGet("comments/{userId}/{eventId}")]
+        [HttpGet("comments/{eventId}")]
         [ProducesResponseType(typeof(List<CommentDto>), 200)]
         [ProducesResponseType(404)]
         [ProducesResponseType(500)]
-        public async Task<IActionResult> GetCommentsOfUserByEvent(string userId, string eventId)
+        public async Task<IActionResult> GetCommentsOfUserByEvent(string eventId)
         {
+            JObject? userInfo = HttpContext.Items["UserInfo"] as JObject;
+
+            if (userInfo == null)
+            {
+                return Unauthorized("User is not logged in.");
+            }
+            
             try
             {
+                string userId = userInfo["data"]["user_id"].ToString();
                 var comments = await eventService.GetCommentsOfUserByEventAsync(userId, eventId);
                 if (comments == null)
                 {
@@ -466,14 +515,28 @@ namespace backend.events
         [ProducesResponseType(500)]
         public async Task<IActionResult> CreateComment([FromBody] CreateCommentDto dto)
         {
+            JObject? userInfo = HttpContext.Items["UserInfo"] as JObject;
+
+            if (userInfo == null)
+            {
+                return Unauthorized("User is not logged in.");
+            }
+            
             if (dto == null || string.IsNullOrEmpty(dto.UserId) || string.IsNullOrEmpty(dto.EventId) || string.IsNullOrEmpty(dto.Text) )
             {
                 return BadRequest(new { error = "Invalid comment data." });
             }
+            
+            string userId = userInfo["data"]["user_id"].ToString();
+
+            if (dto.UserId != userId)
+            {
+                return Unauthorized("User is not authorized.");
+            }
 
             try
             {
-                await eventService.CreateCommentAsync(dto.UserId, dto.EventId, dto.Text);
+                await eventService.CreateCommentAsync(dto);
                 return StatusCode(201, new { message = "Comment created successfully." });
             }
             catch (KeyNotFoundException ex)
